+++ lumibot/backtesting/__init__.py
@@ -1,5 +1,5 @@
 from .alpaca_backtesting import AlpacaBacktesting
-from .alpaca_backtesting_new import AlpacaBacktestingNew
+from .alpaca_backtesting import AlpacaBacktesting
 from .alpha_vantage_backtesting import AlphaVantageBacktesting
 from .backtesting_broker import BacktestingBroker
 from .pandas_backtesting import PandasDataBacktesting
